fix: Consolidate cron jobs to meet plan limits
This commit refactors the cron job implementation to consolidate the page tiering logic into the existing `daily-stats` cron job. This is necessary to comply with the Vercel plan limit of two cron jobs.

The key changes include:

1.  **Cron Job Consolidation:**
    -   The page tiering logic, previously in its own cron job, has been moved into the `daily-stats` cron job at `/api/cron/daily-stats`.
    -   The tiering logic is encapsulated in a `runPageTiering` function, which is called after the daily stats are calculated.

2.  **Redundant Code Removal:**
    -   The separate page tiering cron job file (`/api/cron/page-tiering/route.ts`) and its corresponding test file have been deleted.
    -   The `vercel.json` configuration has been updated to remove the entry for the deleted cron job.

3.  **Updated Tests:**
    -   A new test file has been created for the `daily-stats` cron job.
    -   The tests have been updated to verify both the daily stats calculation and the integrated page tiering functionality.
